requirements.txt
Django==6.0.2
djangorestframework==3.17.1
psycopg[binary]==3.2.12
